move: FSD 구조에 맞춰 이동
