--- src/6_shared/constants/index.ts
@@ -1,3 +1,0 @@
-// PUBLIC shared/constants
-
-export { FINANCIAL_INSTITUTIONS, ACCOUNT_TYPE_LABELS } from './financialInstitutions';
